+++ lib/danger/request_sources/bitbucket_server_api.rb
@@ -31,6 +31,10 @@ module Danger
         @username && !@username.empty? && @password && !@password.empty?
       end
 
+      def pull_request(*)
+        fetch_pr_json
+      end
+
       def fetch_pr_json
         uri = URI(pr_api_endpoint)
         fetch_json(uri)